clean, separate, and sort localization code
